[MBUILDCACHE-73] Unify property value conversions in saving and restoration
