+++ src/main/java/org/apache/maven/buildcache/BuildCacheMojosExecutionStrategy.java
@@ -30,7 +30,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.Strings;
 import org.apache.maven.SessionScoped;
 import org.apache.maven.buildcache.artifact.ArtifactRestorationReport;
@@ -409,19 +408,8 @@ public class BuildCacheMojosExecutionStrategy implements MojosExecutionStrategy 
             final String currentValue;
             try {
                 Object value = ReflectionUtils.getValueIncludingSuperclasses(propertyName, mojo);
-
-                if (value instanceof File) {
-                    Path baseDirPath = project.getBasedir().toPath();
-                    Path path = ((File) value).toPath();
-                    currentValue = normalizedPath(path, baseDirPath);
-                } else if (value instanceof Path) {
-                    Path baseDirPath = project.getBasedir().toPath();
-                    currentValue = normalizedPath(((Path) value), baseDirPath);
-                } else if (value != null && value.getClass().isArray()) {
-                    currentValue = ArrayUtils.toString(value);
-                } else {
-                    currentValue = String.valueOf(value);
-                }
+                Path baseDirPath = project.getBasedir().toPath();
+                currentValue = DtoUtils.normalizeValue(value, baseDirPath);
             } catch (IllegalAccessException e) {
                 LOGGER.error("Cannot extract plugin property {} from mojo {}", propertyName, mojo, e);
                 return false;
@@ -445,33 +433,6 @@ public class BuildCacheMojosExecutionStrategy implements MojosExecutionStrategy 
             }
         }
         return true;
-    }
-
-    /**
-     * Best effort to normalize paths from Mojo fields.
-     * - all absolute paths under project root to be relativized for portability
-     * - redundant '..' and '.' to be removed to have consistent views on all paths
-     * - all relative paths are considered portable and should not be touched
-     * - absolute paths outside of project directory could not be deterministically
-     * relativized and not touched
-     */
-    private static String normalizedPath(Path path, Path baseDirPath) {
-        boolean isProjectSubdir = path.isAbsolute() && path.startsWith(baseDirPath);
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug(
-                    "normalizedPath isProjectSubdir {} path '{}' - baseDirPath '{}', path.isAbsolute() {}, path.startsWith(baseDirPath) {}",
-                    isProjectSubdir,
-                    path,
-                    baseDirPath,
-                    path.isAbsolute(),
-                    path.startsWith(baseDirPath));
-        }
-        Path preparedPath = isProjectSubdir ? baseDirPath.relativize(path) : path;
-        String normalizedPath = preparedPath.normalize().toString();
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("normalizedPath '{}' - {} return {}", path, baseDirPath, normalizedPath);
-        }
-        return normalizedPath;
     }
 
     private enum CacheRestorationStatus {
